Move `TokenSelectionStrategyConfig` to `config.py`
